requirements.txt
# Collectors
shodan>=1.31.0
ipwhois>=1.3.0
requests>=2.31.0
ipinfo>=5.0.0

# AI Providers
groq>=0.9.0
anthropic>=0.29.0
openai>=1.30.0

# Utils
python-dotenv>=1.0.0
rich>=13.7.0

beautifulsoup4>=4.12.0
lxml>=5.0.0